+++ DESAFIO/Feiticeiro.java
@@ -50,7 +50,7 @@ public class Feiticeiro implements Habilidades{
     };
         
         @Override
-    public double  golpeEspecial(){
+    public int golpeEspecial(){
         
         String mensagem = "VOCE DEU O GOLPE ESPECIAL";
         System.out.println(mensagem);
